Removed the SotV patches and added more instructions
Forced a VU JIT flush when a new microprogram is uploaded
Made JIT XGKICK timings a bit better
Made MAC flags update every instruction (to be reworked)
The BIOS successfully boots now. Time to get FFX working...

+++ src/core/ee/vu_jit.cpp
@@ -13,7 +13,7 @@ namespace VU_JIT
 
 VU_JIT64 jit64;
 
-int run(VectorUnit *vu)
+uint16_t run(VectorUnit *vu)
 {
     return jit64.run(*vu);
 }
