+++ src/core/ee/vu_jit.hpp
@@ -1,12 +1,13 @@
 #ifndef VU_JIT_HPP
 #define VU_JIT_HPP
+#include <cstdint>
 
 class VectorUnit;
 
 namespace VU_JIT
 {
 
-int run(VectorUnit* vu);
+uint16_t run(VectorUnit* vu);
 void reset();
 
 };
